Error in LimPho
An error (brackets) was found in lim pho
a method to avoid newton;s method was introduced in FLMALC, additional units were checked, all inputs for MALC are now in Broch units

Also, the range for photoasynthesis is criticazl to get the two method of calculating Pmax to overlap

I had done the normalization of Photoperiod limitation wrong, and the erroneous value was skewing the max. It looks somewhat betta now but still the Jan 1 daylenp is broken.