added Nfrond, Changed items in code for farm3D case
Firstly: the erosion rate should be proportional to the frond area/m2 segment, and not total area in a segment, since the mass is per m2 and the net growth is multiplied by the mass to get the flux per m2.

Secondly, the N uptake is way too high, as the local flux is not corrected for surface. This was fixed in FLMALN and FLMALC

The actual Velocity computed by VELOC was tied to FLMALN in process library

A final big change: I eliminated the need for the user to prescribe LinDenMAL and MALS0 based on the farm configuration/density. Instead I ask them to specify Nfrond/m2 and the correction on LinDenMAL and MALS0 is done in the code in FLMALS and MALDIS. I am not entirely confident there are no other repercussions, but the Broch test case was unaffected by these changes.