Trying effect of multiple fronds, high MALS value
Discovered error in MALDIS for situations where the tip and the foot reside in the same segment. This was fixed.

When biomass is too high, the mortality goes to infinity. This I realize the mrtMal value is for area of a single frond, so therefor I need the m2/m2 value and the nFrond value to get the mortality per m2 per MALS value correct.

After increasing the biomass something about the LimDen is off... might have something to do with the waterlevel variations. I need to look at the data from the sum. This error was due to the fraction biomass affecting LinDen. It should not, since there will be times where only a tiy part of the tip enters the segment, and that tip should not be limited in a lesser way by density than the rest of the frond. Thus the limitation is relative to the whole frond area, but the MALS0 is  increased relative to the nFrond

Just a reminder that LimDen and mrt are the two things causing me issues with units

Also, there is some real messiness about which output parameters are per m2, and also which output parameters pertain to the given segment or to the whole column. This needs some careful attention. The user should not have to "find"the biomass in the water column, should they? Right now MALDIS reports things for all segments based on the bottom segment, but not MALS or MALN or MALC... Some checks and changes were made, no thorough check done.

Checks with the same initial conditions as in broch are promising. Increasing the biomass density has a noticable effect on NO3. Impressions of validity of calculation are promising, although visualizing it is hard ad you cannot only look in a single segment, and the observation are has some issues with use due to adding(?)

Next is to find a way to ensure all looks good in 3D. Make a better routine for visualizing this.