Added LinDenCor, fixed FLCanMALS pointer
Added a linear density correction for low biomasses (LinDenCor)
This revealed that as fronds grew into other segments the limitation functions were improved, meaning fronds of the same length but across different numbers of segments would grow differently.
The issue was partly in FLMALN where the area used for uptake was TotAreMAL, when it should be scaled by FrBmMALS. The same problem was present in FlMALC. This is a big oversight!.

Fixed incorrect footdepth for SWGroDir < 0

Incorrect pointer in FLMALC for CanMALS flux

Fixed HARVMAL to always have a number .ge. 0 for all segments

Previous length to mass ratio was wildly incorrect, should be LinDenMAL = 120

WARNING: computer shut down when I was working on something, not sure if it was incorporated, was a small but important change, and I do not recognize it/ see it reflected in github. Likely the change was lost when the computer shut off?