LinDenCor intercept, stretch test case added
Previous implementation of LinDenCor caused the frond to begin at the correct length but then increase length at an unrealistically high rate. This is because it was scaling the change in length to the change in growth. Instead it should act as an intercept. LinDenCor is now a switch [0 or 1] and calculated LinDenFact in the code. This intercept is determined by the initial biomass condition and is only calculated at the beginning.

Also, FIRST was never switched to false in MALDIS, this switch is now added.

Finally, TotN and TotC were only being calculated if there was growth. Now they are always calculated.

stretch test case added