Fix May 6 commit issue: incorrect cannibalization during respiration, D
Error was present in the cannabalization routine. Source of problem was May 6 commit. Needed to add use of carbon storage in the event that LocCUp was <0, and only remove it from structural if MALC <= MALCmin.

Error in FLMALS, whereby the decay was calculated as a fraction of MALS, but MALS was already corrected for local biomass fraction. This would not be noticed in runs other than those where the biomass covered more than one cell in the vertical.