SeedMass: Strategy to deal with sparsely populated segments
Introduced the concept of seed mass to deal with low frond areas under the 1 frond per m2 assumption, where the low area caused no erosion or density limitation, and the whole culture grew as if it were young. Now the code knows roughly how many fronds there are based on a user defined SeedMass (mass per frond at time of seeding), and does the erosion and density limitation based on the expected area per frond.

Previous notion of nFrond is removed, NFrond is now number of fronds in segment. For the first segment where biomass is present in a column the NFrond is calculated once and saved to the bottom segment, from where it is read each time thereafter. Whichever segment is selected first does not matter because MALS is used, not BmLayMALS, as MALS is common across entire column.

NFrond was complicated due to fact it has to be calculated only once for some substances and stored on bottom layer for those columns that have biomass.

An unresolved problem is that MALDIS would need to know the number of fronds to allocate the biomass across to determine the length of each frond. Right now the length of the fronds is the length of a frond if the biomass was allocated with a single frond per m2. This underestimates the length at sparse plotting densities.

Also fixed NaN and Infinity cases resulting from ammonia preference.